requirements.txt
# Dependencias para el Bot de Trading Mejorado
# Instalar con: pip install -r requirements.txt

# API y requests
requests>=2.32.0
aiohttp>=3.13.0

# Análisis de datos
pandas>=2.3.0
numpy>=2.3.0

# Binance API
python-binance>=1.0.30

# Análisis técnico
ta>=0.11.0

# Utilidades
python-dateutil>=2.9.0
pytz>=2025.2
